requirements.txt
fastapi
SQLAlchemy
uvicorn
pydantic
passlib
bcrypt
PyJWT
pytest 
pytest-cov 
httpx